+++ CTNotificationContent/Templates/Carousel/Model/CarouselProperties.swift
@@ -19,4 +19,7 @@ struct CarouselProperties: Decodable {
     let pt_bg: String?
     let pt_title_clr: String?
     let pt_msg_clr: String?
+    let pt_bg_dark: String?
+    let pt_title_clr_dark: String?
+    let pt_msg_clr_dark: String?
 }