+++ CTNotificationContent/Views/CTCaptionedImageView.swift
@@ -8,6 +8,9 @@ struct CaptionedImageViewComponents {
     var bgColor: String = ""
     var captionColor: String = ""
     var subcaptionColor: String = ""
+    var bgColorDark: String = ""
+    var captionColorDark: String = ""
+    var subcaptionColorDark: String = ""
 }
 
 class CTCaptionedImageView : UIView {
@@ -59,9 +62,6 @@ class CTCaptionedImageView : UIView {
         addSubview(captionLabel)
         addSubview(subcaptionLabel)
 
-        backgroundColor = UIColor(hex: components.bgColor)
-        imageView.backgroundColor = UIColor(hex: components.bgColor)
-
         CTUtiltiy.checkImageUrlValid(imageUrl: components.imageUrl) { [weak self] (imageData) in
             DispatchQueue.main.async {
                 if imageData != nil {
@@ -72,8 +72,34 @@ class CTCaptionedImageView : UIView {
         }
         captionLabel.setHTMLText(components.caption)
         subcaptionLabel.setHTMLText(components.subcaption)
-        captionLabel.textColor = UIColor(hex: components.captionColor)
-        subcaptionLabel.textColor = UIColor(hex: components.subcaptionColor)
+        updateInterfaceColors()
+    }
+    
+    @objc public override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+            super.traitCollectionDidChange(previousTraitCollection)
+            
+            // Check if iOS 12+ API is available before using it
+            if #available(iOS 12.0, *) {
+                if traitCollection.userInterfaceStyle != previousTraitCollection?.userInterfaceStyle {
+                    updateInterfaceColors()
+                }
+            }
+    }
+    
+    func updateInterfaceColors() {
+        // Check if device is in dark mode (iOS 12+)
+        let isDarkMode: Bool
+        
+        if #available(iOS 12.0, *) {
+            isDarkMode = traitCollection.userInterfaceStyle == .dark
+        } else {
+            // For iOS versions before 12.0,using light mode colors since dark mode wasn't officially supported
+            isDarkMode = false
+        }
+        
+        imageView.backgroundColor = UIColor(hex: isDarkMode ? components.bgColorDark : components.bgColor)
+        captionLabel.textColor = UIColor(hex: isDarkMode ? components.captionColorDark : components.captionColor)
+        subcaptionLabel.textColor = UIColor(hex: isDarkMode ? components.subcaptionColorDark : components.subcaptionColor)
     }
     
     func setupConstraints() {